dckvapi log (in progress - todo data streaming to api)

+++ dicm2dckv/dicm2dckv/Repertoires.h
@@ -49,6 +49,6 @@ enum repertoireenum{
 
 
 uint32 repertoireidx( uint8_t *valbytes, uint16 vallength );
-NSString *repertoirename( uint32 repertoireidx);
+NSString *repertoirename( uint16 repertoireidx);
 
 #endif /* Repertoires_h */
