+++ dicm2dckv/dicm2dckv/TransferSyntaxes.m
@@ -175,7 +175,7 @@ uint8 tsidx( uint8_t *valbytes, uint16 vallength )
 }
 
 
-NSString *tsname( uint8 tsidx)
+NSString *tsname( uint16 tsidx)
 {
    return [NSString stringWithUTF8String:tsstr[tsidx]];
 }
